+++ sdk/containerregistry/azure-containerregistry/CHANGELOG.md
@@ -5,7 +5,8 @@
 ### Features Added
 
 ### Breaking Changes
-- Suppress the response error with status code `404` in `delete_blob()`. 
+- Suppress the response error with status code `404` in `delete_blob()`.
+- Added to return blob size in bytes in `upload_blob()`.
 
 ### Bugs Fixed